warning alert works for deleting playlist

+++ src/mytunes_group4/gui/MainView/TunesViewController.java
@@ -3,6 +3,7 @@ package mytunes_group4.gui.MainView;
 import java.io.File;
 import java.io.IOException;
 import java.net.URL;
+import java.util.Optional;
 import java.util.ResourceBundle;
 import java.util.logging.Level;
 import java.util.logging.Logger;
@@ -16,7 +17,10 @@ import javafx.fxml.FXMLLoader;
 import javafx.fxml.Initializable;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
+import javafx.scene.control.Alert;
+import javafx.scene.control.Alert.AlertType;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.Label;
 import javafx.scene.control.ListView;
 import javafx.scene.control.SelectionMode;
@@ -103,7 +107,7 @@ public class TunesViewController implements Initializable
     @Override
     public void initialize(URL url, ResourceBundle rb)
     {
-        
+
         //keeps track of what view (songlist and songs in playlist) is selected
         SongsInPlaylist.getSelectionModel().selectedItemProperty().addListener((obs, old, newVal) ->
         {
@@ -149,8 +153,9 @@ public class TunesViewController implements Initializable
 
     /**
      * The song table view setting the title, artist and genre columns
+     *
      * @throws IOException
-     * @throws DalException 
+     * @throws DalException
      */
     private void songTable() throws IOException, DalException
     {
@@ -162,12 +167,12 @@ public class TunesViewController implements Initializable
         songTableView.getColumns().addAll(viewSongTitle, viewSongArtist, viewSongGenre);
     }
 
-    
     /**
      * The playlist table view setting the name of our playlists
+     *
      * @throws IOException
      * @throws DalException
-     * @throws Exception 
+     * @throws Exception
      */
     private void playlistTable() throws IOException, DalException, Exception
     {
@@ -180,7 +185,7 @@ public class TunesViewController implements Initializable
 
     /*
     Set and show stage addPlaylist
-    */
+     */
     @FXML
     private void addNewPlaylist(ActionEvent event)
     {
@@ -200,7 +205,7 @@ public class TunesViewController implements Initializable
 
     /*
     Set and show stage editPlaylist
-    */
+     */
     @FXML
     private void editPlaylist(ActionEvent event)
     {
@@ -212,10 +217,10 @@ public class TunesViewController implements Initializable
             stage.setScene(new Scene(root1));
             stage.show();
             stage.setAlwaysOnTop(true);
-            
+
             Playlist playlistEdit = playlistTableView.getSelectionModel().getSelectedItem();
             getPlaylistId = playlistEdit.getPlaylistId();
-            
+
         } catch (IOException ex)
         {
             Logger.getLogger(TunesViewController.class.getName()).log(Level.SEVERE, null, ex);
@@ -224,17 +229,32 @@ public class TunesViewController implements Initializable
 
     /*
     Delete playlist
-    */
+     */
     @FXML
     private void deletePlaylist(ActionEvent event) throws Exception
     {
-        tModel.setChosenPlaylist(playlistTableView.getSelectionModel().getSelectedItem());
-        tModel.deletePlaylist(tModel.getChosenPlaylist());
+        Alert alert = new Alert(AlertType.CONFIRMATION);
+        alert.setTitle("A Deletion Confirmation");
+        alert.setHeaderText("Are you sure you want to delete:");
+        alert.setContentText(playlistTableView.getSelectionModel().getSelectedItem() + "?");
+
+        Optional<ButtonType> result = alert.showAndWait();
+        if (result.get() == ButtonType.OK)
+        {
+
+            tModel.setChosenPlaylist(playlistTableView.getSelectionModel().getSelectedItem());
+            tModel.deletePlaylist(tModel.getChosenPlaylist());
+
+        } else
+        {
+            alert.close();
+        }
+
     }
 
     /*
     * Deletes a song in a playlist
-    */
+     */
     @FXML
     private void deleteSongInPlaylist(ActionEvent event) throws Exception
     {
@@ -244,7 +264,7 @@ public class TunesViewController implements Initializable
 
     /*
     * Adds song to playlist
-    */
+     */
     @FXML
     private void addSongToPlaylist(ActionEvent event) throws Exception
     {
@@ -254,10 +274,12 @@ public class TunesViewController implements Initializable
 
     /*
     Set and show stage addSong
-    */
+     */
     @FXML
-    private void addSong(ActionEvent event) {
-        try {
+    private void addSong(ActionEvent event)
+    {
+        try
+        {
             FXMLLoader fxmlLoader = new FXMLLoader(getClass().getResource("/mytunes_group4/gui/SongView/AddSongFXML.fxml"));
             Parent root1 = (Parent) fxmlLoader.load();
             Stage stage = new Stage();
@@ -265,7 +287,8 @@ public class TunesViewController implements Initializable
             stage.show();
             stage.setTitle("Add song");
             stage.setAlwaysOnTop(true);
-        } catch (IOException ex) {
+        } catch (IOException ex)
+        {
             ex.printStackTrace();
             Logger.getLogger(TunesViewController.class.getName()).log(Level.SEVERE, null, ex);
         }
@@ -273,7 +296,7 @@ public class TunesViewController implements Initializable
 
     /*
     Set and show stage editSong and store song values in variables
-    */
+     */
     @FXML
     private void editSong(ActionEvent event)
     {
@@ -286,35 +309,38 @@ public class TunesViewController implements Initializable
             stage.show();
             stage.setTitle("Edit song");
             stage.setAlwaysOnTop(true);
-           
+
             Song songEdit = songTableView.getSelectionModel().getSelectedItem();
             getSongId = songEdit.getSongId();
             getTitle = songEdit.getSongName();
             getArtist = songEdit.getArtistName();
             getGenre = songEdit.getGenre();
             getPath = songEdit.getPath();
-            
+
         } catch (IOException ex)
         {
             Logger.getLogger(TunesViewController.class.getName()).log(Level.SEVERE, null, ex);
         }
     }
-    
+
     /*
     Delete song
-    */
+     */
     @FXML
     private void deleteSong(ActionEvent event) throws Exception
     {
-        
+
         Song selectedSong = songTableView.getSelectionModel().getSelectedItem();
-        if (selectedSong != null) {
-            try {
+        if (selectedSong != null)
+        {
+            try
+            {
                 tModel.deleteSong(selectedSong);
-            } catch (IOException ex) {
+            } catch (IOException ex)
+            {
                 Logger.getLogger(TunesViewController.class.getName()).log(Level.SEVERE, null, ex);
             }
-    }
+        }
     }
 
     @FXML
@@ -429,7 +455,7 @@ public class TunesViewController implements Initializable
             }
         });
     }
-    
+
     /**
      * When a playlist is selected, this displays the songs in the playlist
      */
@@ -473,7 +499,8 @@ public class TunesViewController implements Initializable
     }
 
     /**
-     * This gets the path of our songs when selected and is used to play the selected song
+     * This gets the path of our songs when selected and is used to play the
+     * selected song
      */
     private void setMusicPlayerPath()
     {
@@ -512,7 +539,7 @@ public class TunesViewController implements Initializable
 
     /*
     * gets volume
-    */
+     */
     public double getVolume()
     {
         return currentVolume;
@@ -520,7 +547,7 @@ public class TunesViewController implements Initializable
 
     /*
     * sets volume
-    */
+     */
     public void setVolume(double value)
     {
         if (mediaPlayer != null)
@@ -586,6 +613,4 @@ public class TunesViewController implements Initializable
         }
     }
 
-    }
-
-
+}
